requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
CommonsDownloader==0.5.3
gif2numpy==1.3
idna==2.9
kaitaistruct==0.8
mwclient==0.10.0
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.34
pep8==1.7.1
Pillow==7.1.1
pytesseract==0.3.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.9
